% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorders.R
\name{submit_via_cat}
\alias{submit_via_cat}
\alias{create_submit_to_file}
\alias{create_submit_to_db}
\title{Submit events via cat}
\usage{
submit_via_cat(tutorial_id, tutorial_version, user_id, event, data)

create_submit_to_file(fname)

create_submit_to_db(table_name = "higgins",
  password_table_name = "passwords", restart = FALSE)
}
\arguments{
\item{fname}{text string giving the name of the file for storage}

\item{table_name}{string giving the name of the table in your database to use for storing submissions}

\item{password_table_name}{string giving the name of the table in your database containing
student (and grader) login credentials}

\item{restart}{For development purposes only. If `TRUE` empties the submission table.}
}
\description{
A submitted event consists of a list with several text elements

Submit events to a text file in JSON format
}
