% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recorders.R
\name{submit_via_cat}
\alias{submit_via_cat}
\alias{create_submit_to_file}
\alias{create_submit_to_db}
\title{Submit events via cat}
\usage{
submit_via_cat(tutorial_id, tutorial_version, user_id, event, data)

create_submit_to_file(fname)

create_submit_to_db(dbname = "logevents1", host = "web625.webfaction.com",
  port = 5432, user = "logevents1", password = "learnr555",
  table_name = "higgins", password_table_name = "passwords",
  restart = FALSE)
}
\arguments{
\item{restart}{If `TRUE`, empty the table on the database. Be careful!}
}
\description{
A submitted event consists of a list with several text elements
}
